Refactor: 단계 완료시 갱신, 투표완료시 계산 로직 훅 분리
